Se adelanto el modulo de reportes_arabe de manera que ya es capas de registrar los servicios, empleados y el cliente del modulo en su configuracion, ya es totalmente configurable, se termino el modulo de creacion de reportes de manera que ya se pueden crear los reportes semanales correctamente en la base de datos, actualmente se encuantra en proceso el modulo de la vista del reporte seleccionado, la consulta sql ya funciona correctamente.

+++ application/helpers/manipular_archivos_helper.php
@@ -2,12 +2,17 @@
 
 function rmDir_rf($carpeta)
 {
-  foreach(glob($carpeta . "/*") as $archivos_carpeta){             
-    if (is_dir($archivos_carpeta)){
-      rmDir_rf($archivos_carpeta);
-    } else {
-    unlink($archivos_carpeta);
-    }
-  }
-  rmdir($carpeta);
+
+	if( is_dir($carpeta)) {
+		foreach(glob($carpeta . "/*") as $archivos_carpeta){             
+			if (is_dir($archivos_carpeta)){
+				rmDir_rf($archivos_carpeta);
+			} else {
+			unlink($archivos_carpeta);
+			}
+		}
+		
+		rmdir($carpeta);
+
+	}
  }
